+++ Chatterino.Common/Margin.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Chatterino.Common
+{
+    public class Margin
+    {
+        public int Top { get; set; }
+        public int Right { get; set; }
+        public int Bottom { get; set; }
+        public int Left { get; set; }
+
+        public Margin(int top, int right, int bottom, int left)
+        {
+            Top = top;
+            Right = right;
+            Bottom = bottom;
+            Left = left;
+        }
+
+        public Margin()
+        {
+            
+        }
+    }
+}