+++ Chatterino/Controls/EmoteList.cs
@@ -48,11 +48,11 @@ namespace Chatterino.Controls
                         {
                             if (emotes.Key == 0)
                             {
-                                messages.Add(new Message("Twitch Emoteses"));
+                                messages.Add(new Message("Twitch Emotes"));
                             }
                             else
                             {
-                                messages.Add(new Message("Twitch Subscriber Emoteses"));
+                                messages.Add(new Message("Twitch Subscriber Emotes"));
                             }
 
                             messages.Add(new Message(words));
@@ -64,14 +64,14 @@ namespace Chatterino.Controls
                 //{
                 //    List<Word> words = new List<Word>();
 
-                //    foreach (var emote in Emoteses.ChatterinoEmotes.Values)
+                //    foreach (var emote in Emotes.ChatterinoEmotes.Values)
                 //    {
                 //        words.Add(new Word { Type = SpanType.LazyLoadedImage, Value = emote, Value = emote.Value, CopyText = emote.Name, Link = new Link(LinkType.InsertText, emote.Name + " ") });
                 //    }
 
                 //    if (words.Count != 0)
                 //    {
-                //        messages.Add(new Message("Chatterino Global Emoteses"));
+                //        messages.Add(new Message("Chatterino Global Emotes"));
                 //        messages.Add(new Message(words));
                 //    }
                 //}
@@ -88,7 +88,7 @@ namespace Chatterino.Controls
 
                     if (words.Count != 0)
                     {
-                        messages.Add(new Message("BetterTTV Channel Emoteses"));
+                        messages.Add(new Message("BetterTTV Channel Emotes"));
                         messages.Add(new Message(words));
                     }
                 }
@@ -104,7 +104,7 @@ namespace Chatterino.Controls
 
                     if (words.Count != 0)
                     {
-                        messages.Add(new Message("BetterTTV Global Emoteses"));
+                        messages.Add(new Message("BetterTTV Global Emotes"));
                         messages.Add(new Message(words));
                     }
                 }
@@ -121,7 +121,7 @@ namespace Chatterino.Controls
 
                     if (words.Count != 0)
                     {
-                        messages.Add(new Message("FrankerFaceZ Channel Emoteses"));
+                        messages.Add(new Message("FrankerFaceZ Channel Emotes"));
                         messages.Add(new Message(words));
                     }
                 }
@@ -137,7 +137,7 @@ namespace Chatterino.Controls
 
                     if (words.Count != 0)
                     {
-                        messages.Add(new Message("FrankerFaceZ Global Emoteses"));
+                        messages.Add(new Message("FrankerFaceZ Global Emotes"));
                         messages.Add(new Message(words));
                     }
                 }